+++ src/buffer/host_visible_buffer.rs
@@ -3,7 +3,7 @@ use crate::*;
 pub struct HostVisibleBuffer {
     pub vk_buffer: vk::Buffer,
     pub memory: vk::DeviceMemory,
-    pub size: u64,
+    pub size: usize,
     device: ash::Device,
 }
 
@@ -17,7 +17,7 @@ impl Drop for HostVisibleBuffer {
 }
 
 impl HostVisibleBuffer {
-    pub fn new(size: u64, usage: vk::BufferUsageFlags, gpu: &Gpu) -> HostVisibleBuffer {
+    pub fn new(size: usize, usage: vk::BufferUsageFlags, gpu: &Gpu) -> HostVisibleBuffer {
         let (vk_buffer, memory) = super::new_raw_buffer(
             size,
             usage,
@@ -33,14 +33,19 @@ impl HostVisibleBuffer {
         }
     }
 
-    pub fn upload_data<T>(&self, data: &[T], offset: u64, gpu: &Gpu) {
-        let data_size = std::mem::size_of_val(data) as u64;
+    pub fn upload_data<T>(&self, data: &[T], offset: usize, gpu: &Gpu) {
+        let data_size = std::mem::size_of_val(data);
         debug_assert!(self.size >= offset + data_size);
 
         unsafe {
             let data_ptr = gpu
                 .device
-                .map_memory(self.memory, offset, data_size, vk::MemoryMapFlags::empty())
+                .map_memory(
+                    self.memory,
+                    offset as u64,
+                    data_size as u64,
+                    vk::MemoryMapFlags::empty(),
+                )
                 .expect("Failed to map memory.") as *mut T;
 
             data_ptr.copy_from_nonoverlapping(data.as_ptr(), data.len());
